account: implement account manager to support new account requests
The account manager is a subsystem of the server that will be
responsible for management of all CLM accounts on-chain. This initial
version only supports handling external requests to reserve and create
new accounts at the moment, but this will be extended later on to also
support closing and modifying them. Open accounts are also watched for
their confirmation and spend, though the spend path remains unhandled.

The commit stands alone by itself, and is not yet integrated into the
server as we still need to satisfy its dependencies, which we'll do in
the following commits.