agoradb: introduce awareness of per-batch key
In this commit, we introduce a new key, known as the per-batch key, that
will serve as a tweak for the base keys of an account to ensure account
scripts do not reuse the same keys. The initial per-batch key is a
randomly generated key and is incremented by its curve's base point each
time a new cleared batch happens within the system.

The latest known per-batch key is provided to traders during account
creation and serves as the account's initial per-batch key. Note that
an account's per-batch key won't always match the auctioneer's latest
known per-batch key, as an account may not participate in every cleared
batch of the system. This was done to ease recovery of accounts from the
trader's point-of-view, as otherwise they may need to scan through
multiple per-batch keys before arriving at their account script on-chain
if they are not very active within the system.