venue/matching: add initial match making algorithm
In this commit, we add the initial matching making algorithm we aim to
use. This algorithm is a multi-unit algorithm that supports partial
matches. By partial matches we mean that it's ok if a bid only takes
part of the supply of an ask. When matching in-flight, these remainders
are kept track of, and applied to each order. At the end of the matching
session, all the fields tracking the set of unfilled orders for each
order.

Matching as is is also multi-dimensional, as we need to match both on
the fixed rate as well as the lease duration for the coins. The current
algorithm implemented as O(M*N), where M is the number of bids, and N is
the number of asks. This can likely be optimized much more, but at the
end of the day, we only need to be faster than our batch period (target
10 minutes atm).