venue/matching: create new TradingFeeReport struct to tally fees during execution
In this commit, we introduce the concept of a `TradingFeeReport` which
tallies the total fees paid during a batch. This will be used to
generate a diff of the state of an account before and after the batch,
as well as for our accounting purposes. We abstract away the fees we
charge as the auctioneer behind an interface so we can update it easily
in the future.

We also introduce a new type `FixedRatePremium` which we use to
calculate the lump sum that the taker must pay to the maker.  There's a
rather important TODO in this commit w.r.t ensuring that the auctioneer
is able to extract a base fee even then the proportional fee is close to
zero. We may need to enforce that an account can at least pay for the
base fee before submission.