venue/matching: add new UniformPriceCallMarket impl of BatchAuctioneer
In this commit, we introduce the UniformPriceCallMarket. This is the
primary struct that the auctioneer will interact with in order to clear
markets. Orders are to be added to the in-memory cache of the market as
they're submitted, then finally we attempt a clear with the MaybeClear
method once the time period is up.