account+venue/matching: add new TraderRawKey attribute
In this commit, we add a new `TraderRawKey` attribute to the Account
struct that allows callers to obtain the serialized form of an account
key without needing to parse the full pubkey. In several areas, it's
easier (especially in tests) to use the raw key bytes rather than
needing to generate or obtain the struct first.

This change could've gone a bit deeper, but in the spirit of minimalism,
we only update the areas that we need in order to restore the normal
execution time back to the tests in the matching package. Before this
change, we had to generate an actually public key, which really increased
the execution time of the quickcheck tests.