venue/batchtx: add new chainFeeEstimator to calc fee split for batch tx
In this commit, we add a new `chainFeeEstimator` struct which will be
used in order to calculate aspects like the estimated weight of the
batch transaction, as well as the fees that each trader will need to
pay.

For the auctioneer, our goal is that we pay only the excess "signalling"
costs while the traders pay proportionally based on their footprint
within the batch transaction itself.