multi: unpack rpcserver partially into main server
In this commit, we "unpack" the current rpcServer a bit by moving much
of the sub-system initialization into the main server itself. We then
pass in all required sub-systems directly into the rpc server. We also
add a `Start()` method to the main server, and also move the main
control loop of the daemon into the server. Much of the events as is
right now are drive purely by the client, but a set of upcoming PRs will
start to add server initiated events such as the start of a new batch.