auctioneer: create new auctioneer state machine w/ account init logic
In this commit, we create a new state machine: the Auctioneer. This will
be the primary state machine in the system as it will process both
external and internally generally events to progress the auction.

In this commit, we implement the first few states to go from an empty
system state to a system that's ready to accept orders. In order to get
to this state, we implement a resumable state machine to create what we
call the genesis transaction which pays to the auctioneer's master
account output.