agoradb+matching: make JSON deep compare generic
To test the order batch deep equality, we cannot use
reflect.DeepEqual because the order struct contains net.Addrs which
don't support that for some reason. That's why we use the JSON
serialized format to compare equality. This in turn requires us that
the AccountID type which is used as a map key can be marshaled to
text.