venue: add main batch executor state machine
In this commit, we add the main batch executor state machine. This state
machine will be used by the auctioneer to obtain signatures for a final
valid batch execution transaction. The state machine will attempt to
proceed until either the bach is complete or it encounters a temporary
error. A temporary error is an error such as a trader not sending us a
message in time. The caller is then intended to remove the traders in
the temp error from the batch, and re-submit the batch as a whole.