agoradb: store batch transaction along snapshot, add batch finalization
In this commit, we modify the existing set of methods that store a batch
to also store the batch transaction itself. In addition, we add a new
key to store if a batch has been finalized or not. A batch is only
finalized once the batch execution transaction has confirmed on disk.