auctioneer: complete auction state machine
In this commit, we add two new states: `BatchExecutionState` and
`BatchCommitState` along the way to finalizing the auctioneer state
machine.

The finalized state machines uses the order book to keep track of any
new changes to the set of orders, and updates its internal call market
accordingly. Each time we start up, we'll load these orders from disk
before we request a new subscription stream.

Additionally, we'll also attempt to re-broadcast all prior batches that
aren't yet finalized, marking them as finalized once they're confirmed
on disk.

A new top-level ticker has been added which drives much of the new
logic. Similar to the logic in the `venue.BatchExecutor` struct, we'll
ping pong between the match making and execution phases if we need to
kick out traders to continue operation. Once we have a final batch that
is fully valid, we'll move on to our final state where we broadcast the
batch transaction and wait for its confirmation on-disk.