multi: introduce notion of account diffs
This is done to ensure the auctioneer does not update their main account
state when handling a trader-initiated account modification as they have
no control of the transaction backing the account modification, they
just simply provide a signature back to the trader. With this in place,
the auctioneer now stages any account modifications and commits them
once it sees the account's spend notification.