venue: refresh trader state after a successful batch
In this commit, we'll now ensure that after a recce full batch, for each
trader involved in that batch, we update our in-memory state to reflect
the new state post batch application. This ensures that we'll be able to
process the next batch as we have all our state synchronized with what
the client expected and the objective database state at that time.