account: use same key family for master account output and trader scripts
In this commit, we fix an existing bug by ensuring that we use the same
key for our master account output and what the traders use for the
auctioneer's key in their account outputs. Without this fix, then
existing logic would fail to sign for a trader's output during batch
execution, as we would be using a different key than was expected.