venue/matching: always obtain fresh account view for an order
In this commit, we fix an existing bug that would render us unable to
execute multiple batches back to back. Before this change, we would use
the internal pointer within an order to the account that created this
order. This internal value isn't updated once a batch has been applied,
causing us to attempt to build a batch using stale state, that the
client would rightfully reject.

In this commit, we now have the master maker fetch fresh account state
to ensure we're always using the latest version of a trader's account.