venue: use the NEXT batch key when creating our account in the batch tx
In this commit, we fix a bug that would cause us to re-use a master
account output script. Before this commit, the next output we created
for the auctioneer would be using the same batch key as the prior
output. Instead, we need to rotate this forward by one to ensure we'll
be able to spend the output again during batch to back batch execution
scenarios.