rpc: properly set output index for trader's new account output
In this commit, we update the serving code to ensure we always set the
output index for the trader's new account. Along the way, we now need to
re-create the batch transaction (mainly for its indexes) to ensure we
can send the populated value to the trader when it wishes to restore
from a batch snapshot.