venue: use UnitsUnfulfilled directly in batch storer
In this commit, we fix an existing bug that would prevent orders from
being properly marked as being fully executed on disk. The comment
remove here is wrong, in that the set of orders passed is the output of
the matching making wherein, the `UnitsUnfulfilled` value has already be
updated to reflect the current amount of unfilled units. As a result, if
an order was fully executed, the decrement operation would result in
an under flow, producing a very large value, causing it to always be
marked as partially filled.

We also apply a quick fix to the tests to ensure that the orders used
have the expected amount of units unfilled set.