server: don't use GracefulStop
GracefulStop of grpc.Server waits until all requests have been
handled. This blocks if a trader is misbehaving and not terminating
the stream connection after the auctioneer sent the shutdown signal.
We want to forcefully stop to avoid that.

Additionally, the Stop() already closes the listener so there is no
need to close it again. This also removes the need to return an error
in Stop().