clmrpc+client: remove client
As a consequence of the client split, we now fully remove the client
directory. We cannot keep a copy of the proto files in the server
because we start both the auctioneer and trader in the same process in
the itest. Having the same protos with different packages leads to a
panic in the protobuf library.

NOTE: This commit does not compile as it is a pure code delete
action.