account+agoradb: get reservation by trader key
While recovering accounts for the trader, we want to be able to find
out if the trader had a reservation for a specific trader key before
they lost data. For that we need a lookup of the reservations by
trader key instead of token ID. This is pretty inefficient because we
loop through all reservations. But at any given time there should only
be a very limited number of reservations in the database so this should
not have a big impact on performance.