rpcserver: fix de-multiplexing of batch signed message
For most messages, it's fine to just multiply them to the venue
for each subscribed account. The venue will ignore the message if the
account wasn't involved in a batch.
But for the message that contains signatures, the venue will try to
extract a signature for each account it gets a message for. And if some
accounts were not involved in a batch, no such message should be sent to
the venue.