adminrpc: import auctioneer.proto from $GOPATH
To be able to reuse messages from the auctioneer.proto we import it from
the $GOPATH. This should always pull in the version we have referenced
in our go.mod file. We need to add a PR to the client repo if we want to
change the auctioneer.proto anyway so we might also reference it here as
well instead of having a separate copy (which only brings problems if
the target package differs from the actual implementation used).