multi: use DeriveNextKey to allow for multiple auctioneer instances
This ensures that multiple auctioneer instances can be spun up on the
same network backed by the same lnd seed. Previously, this wasn't the
case as the same base auctioneer key would always be used, leading to
the same account script being funded in the chain. This would trip up
our chain notification logic as it can only dispatch a single
notification per request.

This change also affected account reservations, as now the base
auctioneer key locator isn't statically included in the source code, but
rather stored in the database. If an account is being created while the
auctioneer isn't ready, a proper error is now returned indicating so.