ticker_test: give ticker a second chance to tick
When checking whther the ticker's last tick has been updated, there is a
race between getting the updated value and the ticker ticking and
setting the value.

To address this we handle the case where it has been a while since the
ticker value was updated, and sleep for half a timestamptolerance to
allow the ticker to updat its value.