ticker_test: relax timer accuracy
When checking the NextTickIn value after sleeping, there is a
possibility that the timer ticks before we are able to fetch this value.
In that case the NextTickIn value will be much larger than expected,
causing the test to be flaky.

To address this we increase the timer interval and tolerance, allowing
more slack and decreasing the chance of this happening.