monitoring/dashboards: don't specify an explicit datasource
In this commit, we make a small change to the current dashboard JSON to
allow them to be imported into the Grafana instance running in the
cluster. Rather than specify an explicit datasource (which we never
created in the cluster's Grafana settings?), we need to specify a `null`
source for things to be imported correctly.