matchmaker: filter orders on max fee rate during matchmaking
Using the fee rate we want to use to use for the batch transaction, we
filter out orders that have their max batch fee rate set lower than
this. This lets traders opt out of high fee batches, as they might not
have the funds to pay the on chain fees in this scenario.