venue: allow account to be in StatePendingBatch for match
Because it is not safe to use an account for a batch if there
potentially are "foreign" inputs which we didn't co-sign, we can't allow
any other states for batch participation. Non-co-signed inputs that
could potentially be double spent can happen with an account deposit for
example. And even with an account withdrawal we could run into fee
issues if that update transaction is not yet confirmed and has a very
low fee. That's why we only allow a pending batch to be an unconfirmed
input.