auctioneer test: return non-empty OrderBatch from mock MaybeClear
This commit makes the mock callMarket behave more similar to the real
one, by returning actual matches from MaybeClear. This test change would
have caught that we cleared this orders from the order book and never
re-added them if the batch execution failed.

In addition we add some extra assertions about the submitted and
finalized batches, to make sure we are working with the same batch
all the way through the test.