monitoring: export order fixed rate as distinct label for make shift depth chart
In this commit, we attempt to create a makeshift depth chart on Grafana
using a new label for the order rate. We can group the returned data
based on the order rate, then sum the number of units for each rate to
create a bar graph.

This gets us about 70% of where we want to be, but it seems that
Prometheus doesn't support sorting the x-axis out of the box. Instead,
each time I refresh the graph the bars appear in a different order.
However, the current rendering of this chart at least lets us take a
look at the relative order distribution.