matching: add node conflict predicate
As a preparation for respecting the preference of traders to not match
channels with certain peers and as a way to mitigate recurring
connection or funding problems with some peers, we introduce a predicate
that keeps track of so called node conflicts.
The conflicts can be anything, really. But as soon as a conflict exists
between two nodes, no orders for those nodes will be matched anymore if
this new predicate is used.