itest: fix logger, ensure auction state faster
This commit fixes a few smaller things in the itest. First, the logger
of the etcd server is silenced by bumping its log level to "error" as it
created a lot of log spam otherwise. Second, we speed up the batch
execution test considerably by querying the auction state and then
asserting the mempool instead of always waiting a given time to make
sure the mempool remains empty. Lastly, we fix a bug that left the
mempool filled with a batch TX from a previously added test.