venue+auctioneer: split partial and full reject tracking
This fixes an error with full rejects that would lead to the order book
not shrinking, and the executor getting stuck in an execution loop.

Since a full reject was signaled by adding the trader's own orders to
the OrderRejectMap, the auctioneer actually fail to remove any orders
from the order book. Reason being that it would interpret the orders as
being from the _other party_ in the match, and it would skip removal of
the order since the matched node it would find was not the trader
itself.

We fix this by signaling the full reject individually, and in that case
simply remove all of the trader's orders.