ratings: add new ratings module with node rating abstractions
In this commit, we add a new ratings package that carries a series of
abstractions that we'll use to implement the node ratings system within
subasta. The default implementations are a normal in-memory ratings
database, then a version that backs that simple DB with ratings fetched
from the bos score endpoint. The simple memory version also has support
for a write-thru cache which ideally writes to persistent disk. This
allows us to restart the server but not lose any ratings that we might
have bumped manually in the future.