venue/batchtx: rename to AcctInput to BatchInput
Since we'll soon make it possible to have arbitrary inputs on the batch
transaction, we rename the Acctinput struct to BatchInput. This will
later be needed to look up the information we'll need to sign for the
extra inputs.

We also re-arrange the indices a bit, such that we'll track the input
information for all inputs. Previously the auctioneer's input was
handled separately, now we track it as any other input to the batch via
the batchInputIndex.