rejects: fix reject handling in case of multiple partial matches
This commit fixes a bug in the reject handling, where we would
incorrectly assume the trader was sending incorrect data when there was
multiple partial matches with the same order.

The reason for this was that we would only look at the first match with
the reported order, and when the reporter's key did not match with the
account on this match, we would remove all orders for this node.

A test for this scenario is added.