admin+adminrpc: display order state
We add a new type for ServerAsk and ServerBid to the adminrpc package
because the corresponding types in the poolrpc package don't contain the
order state (as we don't need/want to expose that in the the messages
where those types are used).
To help us more easily debug the state of an order on the server, we
want to know the full state (instead of just archived yes/no).