matching+subastadb: add sub batches and price map
All future batches will persist the clearing prices in a map instead of
a single value because we'll create distinct sub batches with their own
clearing prices each. With this commit we add the new map and use it in
serialization/deserialization.
To keep the commits small, we don't yet remove the old single value
because other packages still use it.