mod+adminrpc: update to latest protobuf version
We want to use docker for compiling the protos. But the version of
protoc available in current images requires a minimum version of the
protobuf library to be used. This only affects the generated code and
not the wire format of the gRPC messages so should have no effect on the
server/client compatibility.
The new protoc version also requires us to set the go package option for
every file so we add that as well.