itest+config: remove autocert and insecure option
Because we always use subasta behind aperture, we don't ever use the
autocert functionality. And since we want all traffic to always be
encrypted between aperture and subasta, we also don't enable the
insecure option on the cluster. Therefore we might as well remove those
options to make the certificate setup a bit simpler.