account+subastadb: fix returning multiple accounts
Because the Store.Accounts() method uses a prefix query to get all
accounts, any diffs were also included in the list. This could've lead
to the same account being tracked by the manager in two different states
if we started the auctioneer while some accounts had a non-committed
diff present.
We fix this issue by never returning account diffs in the Accounts()
method but explicitly loading the diff for each account when resuming it
on startup.