monitoring: add all metrics for archived orders too
We skipped some metrics for archived orders with the goal of reducing
the number of lines in the output. We should be fine with adding all of
them back, performance and storage wise.
The business value of the user agent and unit/duration statistics for
past orders should be quite significant.