matching: add predicate for self chan balance version
We add a matching predicate that makes sure the asker's node is
up-to-date and knows how to account for the self channel
balance feature of bid orders. Old asker nodes wouldn't correctly calculate
an account's diff (and also not open the channel correctly) if they didn't
know of the self channel balance feature. Therefore we need to make sure we
exclude old asker clients from being matched with bid orders that have a self
channel balance set.