matchmaker_test: set correct duration for orders
Since the used duration was not part of the default markets, the market
would never clear, and we would see only negative scenarios.

Because of this one static order, we know that the market should always
clear, so we also alter the test to fail if no market can be cleared.