matchmaker: remove orders below min units left from orderbook
This fixes an issue that would arise from orders getting matched and
ending up with the remainding units below their min units match size.

The orders would be correctly marked as EXECUTED in the DB, but they
would remain in the in-memory orderbook, causing them to potentially be
matched in a following batch, which would be rejected by the trader.