rpcserver+venue: add IsSidecar flags to comms
For the execution of sidecar channels, we'll have two traders online for
the same bid order. One is a normal trader with their own account and
they'll verify and sign the batch itself. The other is possibly a light
client and will only connect out to the taker and wait for the incoming
channel. They don't have a Pool account but instead authenticate
themselves with the channel funding multisig key.
We'll need to interpret the responses from those two clients differently
so we need flags to differentiate between them.