rpcserver: handle sidecar subcription auth correclty
A sidecar recipient is possibly a light client that doesn't have their
own trader account. They'll subscribe to the stream authenticating
themselves with the channel funding multisig key of the channel they'll
receive. We need to handle that case correctly and still create the
necessary structs the venue needs for match making.