venue+rpcserver: handle sidecar orders correctly
Sidecar orders are special in their execution. On the bid side two
traders connect to the auction server. One is the sidecar provider that
has a normal Pool account and uses that account's key as the trader key.
The provider will validate the validity of the batch itself (account
balance diff and so on) and sign the account input itself.
The other trader is the sidecar recipient that uses the bid order's
channel multisig key as the trader key. The recipient creates the
funding shim, waits for the channel to be opened and also sends back the
channel info for the lifetime enforcement.