venue: wait for all channel info to be submitted
Before this commit we only waited for all account witnesses to be
submitted before we would proceed with the batch signing. But since now
there are sidecar acceptor traders that don't submit account witnesses
but the channel information that we need for the lifetime package
enforcement, we also need to make sure we got all of those messages
before we proceed.
This also makes sure we process all messages from all traders, even
those for both participants of a sidecar bid order.