itest: add sidecar reject tests
We need to make sure that we don't proceed with a batch if the sidecar
recipient rejects it. Otherwise we'd create a dead channel if both the
maker and taker agree and sign but the recipient objects.
We create two different tests for the two different phases a reject can
happen:
 1. Prepare phase: If the recipient has the --newnodesonly flag set,
    they don't want new channels from the same node.
 2. Sign phase: If the incoming channel never arrives, we have to retry
    the batch, otherwise the channel will never become active.
    We provoke this by only accepting big channels on the recipient
side.