auctioneerrpc: add new service for synchronous network pipe
In this commit, we add a new service to the auctioneer that implements a
synchronous network pipe. We allow any entity that can pass valid
authentication to create a pipe that can be written to one end, and read
from another. At the time of writing, we only support two authentication
methods: a signature by a valid account pubkey, and a signature of an
existing sidecar ticket.

In the short term, this new sync network pipe will be used to allow
automated negotiation of sidecar channels between a provider and
receiver.

In the future, this network pipe (given a sufficient buffer size) can be
used to get around NATs and act as a sort of network transport wrapper.
In theory, we could also extend the network pipe to be used to allow
pool clients to connect to each other directly, rather than go over the
p2p network.

Beyond that, we plan to use the same implementation to serve a a bridge
between an lnd client running a modified GRUB and an instance of
Terminal running in the web.