hashmail: add new endpoint for stream tear down
In this commit, we add a new endpoint for tearing down an existing
stream. We ensure that only the party that created the stream can tear
it down, and also that they use the exact same authentication mechanism.
An additional test has been added to exercise the new stream deletion
behavior.