hashmail: split out read routine into new stream-local goroutine
In this commit, we split out the logic used to read from the stream into
a new goroutine that's bound to the lifetime of the stream. Before this
change, it was possible that a client connected to the read stream, then
crashed, leaving the _gRPC_ serving goroutine stuck in an io.Reader read
call. This meant that the stream semaphore would never be given up, so
all future connections would show as being occupied.

We fix this by splitting out reading from the stream into a new
goroutine that will send the read bytes over an _unbuffered_ channel to
the main ReadStream serving goroutine. This change means that the main
serving rotouine can come and go, each time always releasing the read
sempahore, and if it fails to fully deliver a given message, then
application retransmission logic can safely be executed.