rpcserver: use defer for disconnecting trader
To avoid a trader still being registered in the active traders map if an
unexpected error occurs, we always disconnect it once the event loop for
it exits.
This should prevent the "already connected, only one stream per trader
allowed" error for traders that fail to authenticate the first time for
some reason.