subastadb: fix possbile inconsistency when updating an account
This commit attempts to correct the STM based account update. When
updating an account if the STM retried we still returned the original
account mutated by our modifiers instead of the actual account as
modified in the DB. We now returns the actual account as present in the
DB to avoid such (unlikely but possible) inconsitencies.