ratings: always reset the scrape time for the bos score scraper
In this commit, we fix a bug that would cause us to never update the set
of scores if we failed parsing or accessing the endpoint the first time
around. I ran into this when updating the subasta lnd nodes: the restart
of the main node caused the endpoint to be unavailable for a period of
time.

Without this commit, if we encounter an error during scraping, then
we'll never actually retry until the system has been restarted.