rpcserver: Make check + add trader operations atomic
Instead of lock and unlock the `connectedStreamsMutex` twice (one for
checking if the trader is already connected and another one to add the
trader to the map of connected traders) do it all in the same lock and
unlock operation.

In this way we remove the possibility of having weird concurrency bugs
triggered by checking twice before adding the trader to the set.