itest: delete testBatchSponsor
Until now, users were able to `sponsor` batches by renew their account
while it was in the `StatePendingBatch`. Renewing an account that is in
`StatePendingBatch` is not supported anymore since the account may
already have extended its expiry height (account autorenew).

If we want to keep the `sponsor` option, we will have to add its own
command.