auctioneer+venue: fix expiry of fully spent accounts
If we set the NewExpiry field on an account diff, the client is going to
look for a pk script with the updated expiry. But for an account that is
fully spent by a batch, the pk script won't change and the client won't
recognize the spend. So we fix this here manually (but will add a fix on
the client side as well, this is only for people that might be running
Pool master).