sidecar: handle sidecar partial rejects
The sidecar acceptor was not reporting back partial rejects properly.
After fixing the pool client, the server needed to be updated to
include predicate conflicts between ask nodes and the sidecar
recipients. Given that we do not have the recipient nodeKey, we
overloaded the `NodeConflictPredicate` to also include conflicts
between `ask.NodeKey` and `bid.MultiSig` for sidecar bids.