chanenforcement: create default package source
`PackageSource` is the interface that takes care of creating,
deleting and enfrocing life time violations for the channels sold.

We were using `EtcdStore` to implement `PackageSource` becuase it
needs to access to the ban data to enforce any violations.

Before migrating to postgres it would be useful to have this logic in
separate packages and leave the `Store` interface only for
retrieving/storing data without any business logic.