server+ratings: make fallback tier configurable
With the intention to simplify the use of subasta in a regtest
environment, we want to make it possible to configure the default node
tier that is returned when no entry is found for a node.
This simplifies order creation because the min node tier parameter does
not have to be specified.