subastadb: add ban sql models
Until now a user could have an entry for a ban or not. If we wanted to
unban a user, we needed to delete the entry, which means that we lost
the historic for that user. Instead of making a ban for a user unique,
let a user has as many as times it has been banned/unbanned so we can
keep track of that user behaviour.