rpc: properly account for the StateExpiredPendingUpdate state
In this commit, we fix an issue that would prevent the public and admin RPC
from returning serialized accounts if one of the accounts was in the state
StateExpiredPendingUpdate.

Back when we added this state
(https://github.com/lightninglabs/subasta/commit/85b025c746862fee3ec64e0c654212fb9cca16f2),
we didn't also update the RPC calls to reflect that this new state could be
returned.