account: disallow reservation for existing account
Fixes an issue that was discovered by a user that attempted to create a
new account after restoring their lnd node from seed and installing a
fresh Pool daemon. This lead to a reservation being created for an
account that already existed from before.
This then lead to problems further down the line when the user attempted
account recovery.
So instead of allowing them to create a reservation for an existing
account, we need to fail hard here to ensure the user runs account
recovery first, before attempting to create new accounts.