account+venue: always send previous outputs
To make subasta combatible with lnd 0.15 in remote signing mode, we need
to make sure to always send the previous outputs, even if they might be
fake ones in some cases. This will be properly addressed in #415 where
we add Taproot support which always requires the previous outputs to be
sent anyway.