subastadb: user order by when useing limit/offset
Until now we are simply setting `LIMIT NULL OFFSET 0` in our queries but
limit should always be used with `ORDER BY` so "pagination" works
properly.

All the columns we are sorting by are indexed with btrees so the queries
do not perform too much extra work. This is something to remember as a
general rule.