rpcserver+venue: move sig parsing
This is a preparation commit that moves the signature parsing logic out
of the RPC server and into the batch executor. Since we'll need to
distinguish between ECDSA and Schnorr signatures based on the account
version, we don't want to do that in the RPC server where we don't have
that information handy. So we forward the raw byte slice signatures to
the executor and parse them there. The actual Schnorr signature parsing
will be done in a later PR.