chanenforcement: fix maturity heigh for script enforced channels
MaturityHeight should hold the absolute heigh for script enforced
channels, as it does for the rest of the system, but it holds a
relative value in chanenforcement. That means that we were not
punishing (banning) users who where closing channels prematurely
even if they were not able to use those coins until the "real"
maturity height was reached.